Snort3/SnortAlerts/requirements.txt
paho-mqtt
scapy